backend/requirements.txt
# ECU Scanner Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyserial>=3.5
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
websockets>=12.0
